requirements.txt
asgiref==3.2.7
astroid==2.4.1
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-cleanup==4.0.0
django-cors-headers==3.3.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-postgres-dropdb==2019.8.24
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.2
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1